Move SimulationInstanceIterator to SimulationInstance
